+++ releases/580_CS80/.gitignore
@@ -2,15 +2,15 @@
 /build/
 /build-*/
 /build_*/
-/experimental-dual-output/build_local_*/
-/experimental-dual-output/build_*/
 
 # Local firmware test drops. Keep release candidates in /uf2/.
 /test-uf2/
 /test_uf2/
 /test-uf2s/
 /test_uf2s/
-/experimental-dual-output/test-uf2/
+
+# Keep local rollback binaries available without committing them.
+/uf2/*rollback*.uf2
 
 # Local scratch ideas that are not part of this card release.
 /FUTURE_PROJECTS.md
