--- releases/580_CS80/experimental-dual-output/tusb_config.h
@@ -1,44 +1,0 @@
-#pragma once
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#define CFG_TUSB_MCU OPT_MCU_RP2040
-#define CFG_TUSB_OS OPT_OS_PICO
-#define CFG_TUSB_RHPORT0_MODE (OPT_MODE_DEVICE | OPT_MODE_HOST)
-
-#ifndef CFG_TUSB_DEBUG
-#define CFG_TUSB_DEBUG 0
-#endif
-
-#define CFG_TUD_ENABLED 1
-#define CFG_TUD_ENDPOINT0_SIZE 64
-
-#define CFG_TUD_CDC 0
-#define CFG_TUD_MSC 0
-#define CFG_TUD_HID 0
-#define CFG_TUD_MIDI 1
-#define CFG_TUD_VENDOR 0
-
-#define CFG_TUD_MIDI_RX_BUFSIZE 256
-#define CFG_TUD_MIDI_TX_BUFSIZE 256
-
-#define CFG_TUH_ENABLED 1
-#define CFG_TUH_DEVICE_MAX 1
-#define CFG_TUH_HUB 1
-#define CFG_TUH_CDC 0
-#define CFG_TUH_HID 0
-#define CFG_TUH_MSC 0
-#define CFG_TUH_VENDOR 0
-#define CFG_TUH_ENDPOINT_MAX 4
-#define CFG_TUH_ENUMERATION_BUFSIZE 256
-#define CFG_TUH_MIDI_RX_BUFSIZE 256
-#define CFG_TUH_MIDI_TX_BUFSIZE 256
-#define CFG_TUH_MIDI_EP_BUFSIZE 64
-#define CFG_TUH_MAX_CABLES 1
-#define CFG_MIDI_HOST_DEVSTRINGS 0
-
-#ifdef __cplusplus
-}
-#endif
